backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
asyncpg>=0.29.0,<1.0.0
anthropic>=0.45.0,<1.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
requests>=2.32.0,<3.0.0
lxml>=6.0.0,<7.0.0
reportlab>=4.4.0,<5.0.0
Pillow>=10.0.0
pydantic>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
validators>=0.22.0,<1.0.0
